refactor IndexModifierType namespace
