+++ src/Dacapo/Domain/Schema/IndexModifier/IndexModifierType/IndexModifierType.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace UcanLab\LaravelDacapo\Dacapo\Domain\Schema\IndexModifier;
+namespace UcanLab\LaravelDacapo\Dacapo\Domain\Schema\IndexModifier\IndexModifierType;
 
 interface IndexModifierType
 {
