+++ src/Dacapo/Domain/Schema/IndexModifier/IndexModifierType/IndexType.php
@@ -2,8 +2,6 @@
 
 namespace UcanLab\LaravelDacapo\Dacapo\Domain\Schema\IndexModifier\IndexModifierType;
 
-use UcanLab\LaravelDacapo\Dacapo\Domain\Schema\IndexModifier\IndexModifierType;
-
 final class IndexType implements IndexModifierType
 {
     /**
